Upd. Integrations. Move wpforms integration to separate class.
