+++ lib/Cleantalk/Antispam/IntegrationsByClass.php
@@ -63,8 +63,7 @@ class IntegrationsByClass
             }
 
             // Public work
-            $skip_integration = $integration->isSkipIntegration();
-            if ($skip_integration) {
+            if ($integration->isSkipIntegration()) {
                 continue;
             }
 
@@ -74,6 +73,13 @@ class IntegrationsByClass
 
     private function isPluginActive($plugin_path)
     {
+        if (is_array($plugin_path)) {
+            foreach ($plugin_path as $path) {
+                if (in_array($path, $this->active_plugins) || in_array($path, $this->active_plugins_wpms)) {
+                    return true;
+                }
+            }
+        }
         return in_array($plugin_path, $this->active_plugins) || in_array($plugin_path, $this->active_plugins_wpms);
     }
 }
